nvme: add low-level APIs to map prp/sgl without @rq
Add low-level helpers to map PRP and SGL data and data list which
contains the actual prp and sgl entries within a memory page.
nvme_rq_map* helpers use datalist as @rq->page.vaddr which is allocated
at the initial time.  Some applications might want to locate the
datalist memory page in another location such as CMB (Controller Memory
Buffer) rather than the host memory.  In that case, application should
update @rq->page, but it's not a good idea since it's a too much
libvfn-specific data structure and it has its own policy.  To give more
flexible options to upper layer application, this patch adds low-level
APIs which are much more generic than nvme_rq_map* helpers by receving a
@prplist and @seg for PRP and SGL list from the caller.

The newly added public helpers are:

    - nvme_map_prp
    - nvme_mapv_prp
    - nvme_mapv_sgl

This commit does not have functional changes of the existing APIs.
